Added Pointer managers, commands and buffer manager

+++ app/src/main/java/com/example/firstapp/interfaces/IPointerRecorder.java
@@ -1,7 +1,7 @@
 package com.example.firstapp.interfaces;
 
 public interface IPointerRecorder {
-    void recordPosition(int x, int y, int z);
+    void recordPosition(double x, double y, double z);
     boolean isDrawing();
     boolean isMoving();
     void setIsDrawing(boolean value);
